Assignment 9: Layered Pattern [수정]
