+++ Inseok_45/models/dataSource.js
@@ -1,0 +1,12 @@
+const {DataSource} = require('typeorm');
+
+const dataSource = new DataSource({
+    type: process.env.TYPEORM_CONNECTION,
+    host: process.env.TYPEORM_HOST,
+    port: process.env.TYPEORM_PORT,
+    username: process.env.TYPEORM_USERNAME,
+    password: process.env.TYPEORM_PASSWORD,
+    database: process.env.TYPEORM_DATABASE
+});
+
+module.exports =  dataSource ;
