+++ Inseok_45/models/userDao.js
@@ -1,22 +1,4 @@
-const { DataSource } = require('typeorm');
-
-const dataSource = new DataSource({
-    type: process.env.TYPEORM_CONNECTION,
-    host: process.env.TYPEORM_HOST,
-    port: process.env.TYPEORM_PORT,
-    username: process.env.TYPEORM_USERNAME,
-    password: process.env.TYPEORM_PASSWORD,
-    database: process.env.TYPEORM_DATABASE
-});
-
-dataSource
-    .initialize()
-    .then(() => {
-        console.log("Data source has been initialized!");
-    }).catch((err) => {
-        console.log("Data source Not Initialize : ", err);
-            dataSource.destroy();
-    });
+const dataSource  = require('./dataSource');
 
 const createUser = async ( firstName, lastName, email, phoneNumber, age, userName, password ) => {
     try {
